RestauraceFelLibraryHibernate/src/main/java/cz/cvut/fel/restauracefel/hibernate/Uzaverka.java
/*
 * To change this template, choose Tools | Templates
 * and open the template in the editor.
 */

package cz.cvut.fel.restauracefel.hibernate;

import java.io.Serializable;
import java.util.Date;

/**
 *
 * @author Lukas Camra
 */
public class Uzaverka implements Serializable {

    private static final long serialVersionUID = -3485621218713955338L;

    private int id;
    private User user;
    private Date date;
    private boolean closed;

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public Date getDate() {
        return date;
    }

    public void setDate(Date date) {
        this.date = date;
    }

    public User getUser() {
        return user;
    }

    public void setUser(User user) {
        this.user = user;
    }

    public boolean isClosed() {
        return closed;
    }

    public void setClosed(boolean closed) {
        this.closed = closed;
    }
}
